# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "selfevals",
    "description": "HTTP bridge for the selfevals evals framework.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/workspaces/{workspace_id}/experiments": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments Index",
        "operationId": "experiments_index_api_workspaces__workspace_id__experiments_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExperimentState"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by experiment state (e.g. running, completed).",
              "title": "State"
            },
            "description": "Filter by experiment state (e.g. running, completed)."
          },
          {
            "name": "feature",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to experiments whose taxonomy.target_features contains this.",
              "title": "Feature"
            },
            "description": "Filter to experiments whose taxonomy.target_features contains this."
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentListPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/run": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments Run",
        "operationId": "experiments_run_api_workspaces__workspace_id__experiments_run_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunExperimentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}/cancel": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments Cancel",
        "operationId": "experiments_cancel_api_workspaces__workspace_id__experiments__experiment_id__cancel_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments Show",
        "operationId": "experiments_show_api_workspaces__workspace_id__experiments__experiment_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}/iterations": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments Iterations",
        "operationId": "experiments_iterations_api_workspaces__workspace_id__experiments__experiment_id__iterations_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IterationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}/cases": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments Cases",
        "operationId": "experiments_cases_api_workspaces__workspace_id__experiments__experiment_id__cases_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}/results": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Per-scenario expected vs detected vs matched (best iteration)",
        "description": "Returns one `ScenarioResult` per case of the best iteration. `expected`/`detected` are **derived per declared dimension**: a case that declares `structured_output` gets only that; a `must_include` case gets substrings + the produced `content` (+ `missing` on a gap); a tool case gets `required_tools` vs `tools_invoked`. Undeclared dimensions are omitted (not null), so the payload stays compact. `message` is the classified reply. A conversation case carries its per-turn breakdown in `turns[]` (same shape) when called with `?include=turns`.\n\n**Migration (0.8.0 \u2192 0.9.0, breaking):** the old flat `CaseResultRow` (with fixed `detected={content,structured_output,tools_invoked}`) is replaced by this recursive, dimension-derived `ScenarioResult`.",
        "operationId": "experiments_results_api_workspaces__workspace_id__experiments__experiment_id__results_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated expansions. `turns` expands each conversation case into per-turn `ScenarioResult`s (off by default \u2014 the case-level grid stays one representative trace per case).",
              "title": "Include"
            },
            "description": "Comma-separated expansions. `turns` expands each conversation case into per-turn `ScenarioResult`s (off by default \u2014 the case-level grid stays one representative trace per case)."
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}/decisions": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments Decisions",
        "operationId": "experiments_decisions_api_workspaces__workspace_id__experiments__experiment_id__decisions_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionRecordResponse"
                  },
                  "title": "Response Experiments Decisions Api Workspaces  Workspace Id  Experiments  Experiment Id  Decisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExperimentListPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExperimentSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "ExperimentListPage",
        "description": "Paginated `GET /workspaces/{ws}/experiments` (A8).\n\nSchema-versioned envelope so the FE can offer \"load more\" /\ncursor-style navigation when an installation accumulates many\nexperiments. Ships before any FE pagination UI lands so the\ncontract is stable in V0 \u2014 the day a user has 500+ experiments\nwe don't need a breaking API change to surface a pager."
      },
...
      "ExperimentState": {
        "type": "string",
        "enum": [
          "draft",
          "queued",
          "running",
          "paused",
          "completed",
          "aborted",
          "superseded"
        ],
        "title": "ExperimentState",
        "description": "Experiment lifecycle states. Operational \u00a7A.2."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}